requirements.txt
Django==1.7
feedparser==5.2.1
django-blog-zinnia==0.15.2
django-contrib-comments==1.6.1
django-mptt==0.7.4
django-scheduler==0.7.2
django-tagging==0.4
django-xmlrpc==0.1.5
python-social-auth==0.2.12
mots-vides==2015.5.11
pyparsing==2.0.3
pytz==2015.4
beautifulsoup4==4.3.2
pillow==2.8.2
